refactor: restructure typings directory and focus on separation of concern
