+++ Source/Typings/Core/Config.d.ts
