+++ Source/Typings/External/Weather.d.ts
